+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/InfrastructureFormula.kt
@@ -6,6 +6,7 @@ import com.atlassian.performance.tools.awsinfrastructure.NetworkFormula
 import com.atlassian.performance.tools.awsinfrastructure.api.jira.DataCenterFormula
 import com.atlassian.performance.tools.awsinfrastructure.api.jira.JiraFormula
 import com.atlassian.performance.tools.awsinfrastructure.api.jira.StandaloneFormula
+import com.atlassian.performance.tools.awsinfrastructure.api.virtualusers.Ec2VirtualUsersFormula
 import com.atlassian.performance.tools.awsinfrastructure.api.virtualusers.StackVirtualUsersFormula
 import com.atlassian.performance.tools.awsinfrastructure.api.virtualusers.VirtualUsersFormula
 import com.atlassian.performance.tools.awsinfrastructure.virtualusers.S3ResultsTransport
@@ -22,6 +23,7 @@ import java.util.concurrent.Executors
  * - [DataCenterFormula]
  * - [StandaloneFormula]
  * - [StackVirtualUsersFormula]
+ * - [Ec2VirtualUsersFormula]
  */
 class InfrastructureFormula<out T : VirtualUsers>(
     private val investment: Investment,
@@ -114,9 +116,10 @@ class InfrastructureFormula<out T : VirtualUsers>(
     private fun overrideVuNetwork(
         network: Network
     ): VirtualUsersFormula<T> = when (virtualUsersFormula) {
-        is StackVirtualUsersFormula -> StackVirtualUsersFormula.Builder(virtualUsersFormula).network(network).build() as VirtualUsersFormula<T>
+        is StackVirtualUsersFormula -> StackVirtualUsersFormula.Builder(virtualUsersFormula).network(network).build()
+        is Ec2VirtualUsersFormula -> Ec2VirtualUsersFormula.Builder(virtualUsersFormula).network(network).build()
         else -> virtualUsersFormula
-    }
+    } as VirtualUsersFormula<T>
 }
 
 class ProvisionedInfrastructure<out T : VirtualUsers>(